dags/tasks/monitor.py
import pandas as pd
import logging
from datetime import datetime
from storage_iceberg import IcebergStorage

class Monitoring:
    def __init__(self, report_file: str = "pipeline_report.csv"):
        """
        Initialize monitoring logger and report file path
        """
        # Steps:
        # 1. Set up logger
        # 2. Initialize CSV file if not exists with headers:
        #    ["run_id","records_processed","anomalies_detected","iceberg_files","partitions_updated","task_times"]
        self.report_file = report_file
        pass

    def log_task_stats(self, run_id: str, task_name: str, start_time: datetime, end_time: datetime):
        """
        Track processing time per DAG task
        """
        # Steps:
        # 1. Calculate duration
        # 2. Store in temporary dict/list for the run
        pass

    def record_metrics(self, run_id: str, records_processed: int, anomalies_detected: int, iceberg_storage: IcebergStorage):
        """
        Collect DAG run metrics
        """
        # Steps:
        # 1. Get Iceberg commit stats from iceberg_storage.get_commit_stats()
        # 2. Compile metrics dict:
        #    {
        #       "run_id": run_id,
        #       "records_processed": records_processed,
        #       "anomalies_detected": anomalies_detected,
        #       "iceberg_files": ...,
        #       "partitions_updated": ...,
        #       "task_times": ...
        #    }
        pass

    def write_report(self):
        """
        Append current run metrics to CSV
        """
        # Steps:
        # 1. Convert metrics to DataFrame
        # 2. Append to report CSV
        pass
